Rework connection module
Make connection type abstract, and make an unique reference for each connection,
so that the client can match on messages from that connection (retrieve the
connection reference using pipesock_conn:get_ref/1).

In addition to that, the client can now specify connection options in the form
of a map, containing:

  id_len => non_neg_integer()
    Determines the number bits used for the message header (used to distinguish
    unique messages). If not specified, the default of 16 bits is used.

    Care must be placed so that the number of clients sharing a single
    connection doesn't go over 2^id_len.

  buf_watermark => non_neg_integer()
    The max number of messages sitting in the connection buffer. Once it goes
    over this number, the connection will automatically flush it.

  cork_len => non_neg_integer()
    The max number of milis between buffer flushes.

Whenever cork_len ms, or buf_watermark messages are buffered, whatever happens
first, the connection will flush the buffer and send all the messages.